requirements.txt
fastapi
uvicorn
python-dotenv
twilio
elevenlabs
gspread
google-auth
pydantic
ngrok